fix: Add profile header spacing and update survey link
- Add margin-top to profile header for better spacing
- Update custom generator survey link to new Tally form
- Improve knowledge selector integration in text generators
- Remove deprecated KnowledgeSelector component
