+++ gruenerator_frontend/src/assets/styles/features/auth/profile.css
@@ -9,6 +9,7 @@
 
 
 .profile-header {
+  margin-top: var(--spacing-large);
   margin-bottom: var(--spacing-large);
   text-align: center;
 }
