--- gruenerator_frontend/src/components/common/KnowledgeSelector/KnowledgeSelector.jsx
@@ -1,607 +1,0 @@
-import React, { useState, useEffect, useMemo, useCallback, useRef } from 'react';
-import PropTypes from 'prop-types';
-import EnhancedSelect from '../EnhancedSelect';
-import '../../../assets/styles/components/ui/knowledge-selector.css';
-import { useGeneratorKnowledgeStore } from '../../../stores/core/generatorKnowledgeStore';
-import { useGroups, useAllGroupsContent } from '../../../features/groups/hooks/useGroups';
-import { useBetaFeatures } from '../../../hooks/useBetaFeatures';
-import { useDocumentsStore } from '../../../stores/documentsStore';
-import { useAuth } from '../../../hooks/useAuth';
-import { useInstructionsStatusForType } from '../../../features/auth/hooks/useInstructionsStatus';
-
-
-/**
- * ProfileSelector - Separate component for custom instructions only
- */
-const ProfileSelector = ({ disabled = false, tabIndex, instructionType = 'social', disableMenuPortal = false }) => {
-  const { 
-    source, 
-    setSource
-  } = useGeneratorKnowledgeStore();
-  
-  // Get groups and beta features for source selection
-  const { isLoading: isLoadingBetaFeatures } = useBetaFeatures();
-  const anweisungenBetaEnabled = true;
-  const { 
-    userGroups: groups, 
-    isLoadingGroups, 
-    errorGroups: groupsError 
-  } = useGroups();
-  
-  // Handle profile source selection (for instructions only)
-  const handleSourceChange = useCallback((selectedOption) => {
-    if (!selectedOption) return;
-
-    const value = selectedOption.value;
-
-    if (value === 'neutral') {
-      setSource({ type: 'neutral', id: null, name: null });
-    } else if (value === 'user') {
-      setSource({ type: 'user', id: null, name: 'Meine Anweisungen' });
-    } else if (value.startsWith('group-')) {
-      const groupId = value.substring("group-".length);
-      const selectedGroup = groups.find(g => g.id === groupId);
-      if (selectedGroup) {
-        setSource({ type: 'group', id: selectedGroup.id, name: selectedGroup.name });
-      }
-    }
-  }, [groups, setSource]);
-
-  // Get current source value for the select
-  const getCurrentSourceValue = useCallback(() => {
-    if (source.type === 'neutral') return 'neutral';
-    if (source.type === 'user') return 'user';
-    if (source.type === 'group') return `group-${source.id}`;
-    return 'neutral';
-  }, [source.type, source.id]);
-
-  // Memoize source options array
-  const sourceOptions = useMemo(() => {
-    const baseOptions = [
-      { value: 'neutral', label: 'Keine Anweisungen' },
-      { value: 'user', label: 'Meine Anweisungen' }
-    ];
-
-    const loadingOptions = [];
-    if (isLoadingBetaFeatures) {
-      loadingOptions.push({ value: '', label: 'Lade Beta Features...', disabled: true });
-    }
-    if (isLoadingGroups) {
-      loadingOptions.push({ value: '', label: 'Lade Gruppen...', disabled: true });
-    }
-    if (groupsError && !isLoadingGroups) {
-      loadingOptions.push({ value: '', label: 'Fehler beim Laden der Gruppen', disabled: true });
-    }
-
-    const groupOptions = (groups && !isLoadingGroups && !groupsError) 
-      ? groups.map(group => ({
-          value: `group-${group.id}`,
-          label: `${group.name} Profil`
-        }))
-      : [];
-
-    return [...baseOptions, ...loadingOptions, ...groupOptions];
-  }, [groups, isLoadingGroups, isLoadingBetaFeatures, groupsError]);
-  
-  return (
-    <div className="profile-selector">
-      <EnhancedSelect
-        inputId="profile-source"
-        classNamePrefix="enhanced-knowledge-select"
-        className="enhanced-knowledge-select"
-        label="Profil für Anweisungen auswählen"
-        options={sourceOptions}
-        value={sourceOptions.find(opt => opt.value === getCurrentSourceValue())}
-        onChange={handleSourceChange}
-        isDisabled={isLoadingGroups || isLoadingBetaFeatures || !anweisungenBetaEnabled || disabled}
-        placeholder="Auswählen"
-        tabIndex={tabIndex}
-        isMulti={false}
-        isClearable={false}
-        isSearchable={false}
-        closeMenuOnSelect={true}
-        hideSelectedOptions={false}
-        blurInputOnSelect={true}
-        openMenuOnFocus={false}
-        tabSelectsValue={true}
-        captureMenuScroll={false}
-        menuShouldBlockScroll={false}
-        menuShouldScrollIntoView={false}
-        menuPortalTarget={disableMenuPortal ? null : document.body}
-        menuPosition={disableMenuPortal ? "absolute" : "fixed"}
-      />
-    </div>
-  );
-};
-
-ProfileSelector.propTypes = {
-  disabled: PropTypes.bool,
-  tabIndex: PropTypes.number,
-  instructionType: PropTypes.oneOf(['antrag', 'social', 'universal', 'gruenejugend']),
-  disableMenuPortal: PropTypes.bool
-};
-
-/**
- * EnhancedKnowledgeSelector - Unified knowledge selector for all sources with React Portal
- */
-const EnhancedKnowledgeSelector = ({
-  onKnowledgeSelection,
-  disabled = false,
-  tabIndex,
-  disableMenuPortal = false
-}) => {
-  const {
-    availableKnowledge,
-    selectedKnowledgeIds,
-    toggleSelection,
-    // Document state
-    selectedDocumentIds,
-    toggleDocumentSelection,
-    isExtractingDocumentContent,
-    documentExtractionInfo,
-    isLoadingDocuments,
-    // Text state
-    availableTexts,
-    selectedTextIds,
-    isLoadingTexts,
-    toggleTextSelection,
-    fetchTexts,
-    // UI Configuration
-    uiConfig
-  } = useGeneratorKnowledgeStore();
-
-  // Extract UI config values for cleaner code
-  const {
-    enableKnowledge = false,
-    enableDocuments = false,
-    enableTexts = false
-  } = uiConfig;
-
-  // Get user groups and authentication
-  const {
-    userGroups: groups
-  } = useGroups();
-
-  const { user } = useAuth();
-  
-  // Load content from all groups using the new hook
-  const {
-    allGroupContent,
-    groupContentErrors,
-    hasGroupErrors,
-    isLoadingAllGroupsContent,
-    isErrorAllGroupsContent,
-    errorAllGroupsContent
-  } = useAllGroupsContent({ 
-    isActive: true,
-    enabled: true // Always load group content
-  });
-  
-  // For backwards compatibility, alias the loading state
-  const isLoadingAllGroups = isLoadingAllGroupsContent;
-
-  // Derived state: Combined loading state for all content sources
-  const isLoadingAnyContent = useMemo(() => {
-    return isLoadingAllGroups || isLoadingTexts || isLoadingDocuments;
-  }, [isLoadingAllGroups, isLoadingTexts, isLoadingDocuments]);
-
-  // Removed console.log for performance
-
-  // Get documents from generatorKnowledgeStore (now properly synced by useKnowledge)
-  const {
-    availableDocuments: documentsFromKnowledgeStore
-  } = useGeneratorKnowledgeStore();
-
-  // Use documents from the knowledge store which are synced from documentsStore
-  const documentsStoreData = documentsFromKnowledgeStore;
-
-  // Note: Document fetching is now handled by useKnowledge hook in parent components
-  // This prevents duplicate fetches and ensures proper synchronization
-
-  useEffect(() => {
-    if (enableTexts) {
-      fetchTexts();
-    }
-  }, [enableTexts, fetchTexts]);
-
-  // State for current search term to trigger re-sorting
-  const [currentSearchTerm, setCurrentSearchTerm] = useState('');
-
-  // Helper function to truncate long titles
-  const truncateTitle = useCallback((title, maxLength = 80) => {
-    if (!title || title.length <= maxLength) return title;
-    return title.substring(0, maxLength) + '...';
-  }, []);
-
-  // Relevance scoring function for search ranking
-  const calculateRelevanceScore = useCallback((option, searchTerm) => {
-    if (!searchTerm) return 0;
-    
-    const title = option.label.toLowerCase();
-    const content = option.searchableContent || '';
-    const searchTermLower = searchTerm.toLowerCase();
-    
-    let score = 0;
-    
-    // 1. Exact title match (highest priority)
-    if (title === searchTermLower) {
-      score += 100;
-    }
-    // 2. Title starts with search term
-    else if (title.startsWith(searchTermLower)) {
-      score += 80;
-    }
-    // 3. Title contains search term
-    else if (title.includes(searchTermLower)) {
-      score += 50;
-    }
-    
-    // 4. Content frequency scoring
-    const occurrences = (content.match(new RegExp(searchTermLower, 'g')) || []).length;
-    score += Math.min(occurrences * 5, 25); // Max 25 points for frequency
-    
-    // 5. Position scoring (earlier occurrence = higher score)
-    const firstIndex = content.indexOf(searchTermLower);
-    if (firstIndex !== -1) {
-      const positionScore = Math.max(10 - (firstIndex / content.length) * 10, 1);
-      score += positionScore;
-    }
-    
-    // 6. Content type priority
-    if (option.itemType === 'knowledge') score += 10;
-    else if (option.itemType === 'text') score += 5;
-    // documents get no bonus (lowest priority)
-    
-    // 7. Multi-word search support
-    const searchWords = searchTermLower.split(' ').filter(word => word.length > 1);
-    if (searchWords.length > 1) {
-      const wordMatches = searchWords.filter(word => content.includes(word)).length;
-      score += (wordMatches / searchWords.length) * 15; // Bonus for multi-word matches
-    }
-    
-    return score;
-  }, []);
-
-  // Helper function to highlight search terms in text
-  const highlightSearchTerm = useCallback((text, searchTerm) => {
-    if (!searchTerm || !searchTerm.trim()) return text;
-    
-    const terms = searchTerm.toLowerCase().split(' ').filter(term => term.length > 0);
-    let highlightedText = text;
-    
-    terms.forEach(term => {
-      const regex = new RegExp(`(${term})`, 'gi');
-      highlightedText = highlightedText.replace(regex, '<mark>$1</mark>');
-    });
-    
-    return highlightedText;
-  }, []);
-
-  // Unified options from all sources (user + all groups)
-  const allKnowledgeOptions = useMemo(() => {
-    const allOptions = [];
-    
-    // User knowledge
-    if (enableKnowledge && availableKnowledge.length > 0) {
-      const userKnowledgeItems = availableKnowledge.map(item => ({
-        value: `knowledge_${item.id}`,
-        label: truncateTitle(item.title),
-        iconType: item.type || 'knowledge',
-        tag: { label: 'Mein Profil', variant: 'user' },
-        itemType: 'knowledge',
-        originalId: item.id,
-        sourceType: 'user',
-        searchableContent: `${item.title} ${item.content || ''}`.toLowerCase(),
-        created_at: item.created_at || null
-      }));
-      allOptions.push(...userKnowledgeItems);
-    }
-    
-    // User documents
-    if (enableDocuments && documentsStoreData.length > 0) {
-      const userDocumentItems = documentsStoreData
-        .filter(doc => doc.status === 'completed')
-        .map(doc => ({
-          value: `document_${doc.id}`,
-          label: truncateTitle(doc.title),
-          iconType: 'user_document',
-          tag: { label: 'Mein Profil', variant: 'user' },
-          itemType: 'document',
-          originalId: doc.id,
-          sourceType: 'user',
-          searchableContent: `${doc.title} ${doc.filename || ''} ${doc.ocr_text || ''}`.toLowerCase(),
-          created_at: doc.created_at || null
-        }));
-      allOptions.push(...userDocumentItems);
-    }
-    
-    // User texts
-    if (enableTexts && availableTexts.length > 0) {
-      const userTextItems = availableTexts.map(text => ({
-        value: `text_${text.id}`,
-        label: truncateTitle(text.title),
-        iconType: 'user_text',
-        tag: { label: 'Mein Profil', variant: 'user' },
-        itemType: 'text',
-        originalId: text.id,
-        sourceType: 'user',
-        searchableContent: `${text.title} ${text.type || ''} ${text.full_content || text.content || ''}`.toLowerCase(),
-        created_at: text.created_at || null
-      }));
-      allOptions.push(...userTextItems);
-    }
-    
-    // Group content from all groups
-    if (allGroupContent.length > 0) {
-      const groupItems = allGroupContent.map(item => {
-        const baseItem = {
-          originalId: item.id,
-          sourceType: 'group',
-          tag: { label: item.groupName, variant: 'group' },
-          created_at: item.created_at || null
-        };
-        
-        // Determine item type and create appropriate option
-        if (item.type === 'knowledge' || (!item.filename && !item.full_content)) {
-          return {
-            ...baseItem,
-            value: `group_knowledge_${item.id}`,
-            label: truncateTitle(item.title),
-            iconType: 'group_knowledge',
-            itemType: 'knowledge',
-            searchableContent: `${item.title} ${item.content || ''}`.toLowerCase()
-          };
-        } else if (item.filename || item.ocr_text) {
-          return {
-            ...baseItem,
-            value: `group_document_${item.id}`,
-            label: truncateTitle(item.title),
-            iconType: 'group_document',
-            itemType: 'document',
-            searchableContent: `${item.title} ${item.filename || ''} ${item.ocr_text || ''}`.toLowerCase()
-          };
-        } else {
-          return {
-            ...baseItem,
-            value: `group_text_${item.id}`,
-            label: truncateTitle(item.title),
-            iconType: 'group_text',
-            itemType: 'text',
-            searchableContent: `${item.title} ${item.type || ''} ${item.full_content || item.content || ''}`.toLowerCase()
-          };
-        }
-      });
-      allOptions.push(...groupItems);
-    }
-    
-    return allOptions;
-  }, [enableKnowledge, enableDocuments, enableTexts, availableKnowledge, documentsStoreData, availableTexts, allGroupContent, truncateTitle]);
-
-  // Sorted and filtered options based on current search term
-  const knowledgeOptions = useMemo(() => {
-    if (!currentSearchTerm || currentSearchTerm.trim() === '') {
-      // No search term - return all options sorted with knowledge items first
-      return allKnowledgeOptions.sort((a, b) => {
-        // Primary sort: knowledge items first (user knowledge, then group knowledge)
-        const getPriority = (item) => {
-          if (item.sourceType === 'user' && item.itemType === 'knowledge') return 0;
-          if (item.sourceType === 'group' && item.itemType === 'knowledge') return 1;
-          if (item.sourceType === 'user' && item.itemType === 'text') return 2;
-          if (item.sourceType === 'group' && item.itemType === 'text') return 3;
-          if (item.sourceType === 'user' && item.itemType === 'document') return 4;
-          if (item.sourceType === 'group' && item.itemType === 'document') return 5;
-          return 6; // fallback
-        };
-        
-        const priorityDiff = getPriority(a) - getPriority(b);
-        if (priorityDiff !== 0) return priorityDiff;
-        
-        // Secondary sort: recency (newer first)
-        const aDate = new Date(a.created_at || 0);
-        const bDate = new Date(b.created_at || 0);
-        return bDate - aDate;
-      });
-    }
-    
-    const searchTerm = currentSearchTerm.trim();
-    
-    // Score and filter options
-    const scoredOptions = allKnowledgeOptions
-      .map(option => ({
-        ...option,
-        relevanceScore: calculateRelevanceScore(option, searchTerm)
-      }))
-      .filter(option => option.relevanceScore > 0) // Only show items with matches
-      .sort((a, b) => b.relevanceScore - a.relevanceScore); // Sort by relevance score (highest first)
-    
-    return scoredOptions;
-  }, [allKnowledgeOptions, currentSearchTerm, calculateRelevanceScore]);
-
-
-  const handleKnowledgeChange = useCallback((selectedOptions) => {
-    const newSelectedValues = selectedOptions ? selectedOptions.map(option => option.value) : [];
-    
-    // Separate different types of selections (including group content)
-    const newKnowledgeIds = newSelectedValues
-      .filter(value => value.startsWith('knowledge_') || value.startsWith('group_knowledge_'))
-      .map(value => value.replace(/^(knowledge_|group_knowledge_)/, ''));
-    const newDocumentIds = newSelectedValues
-      .filter(value => value.startsWith('document_') || value.startsWith('group_document_'))
-      .map(value => value.replace(/^(document_|group_document_)/, ''));
-    const newTextIds = newSelectedValues
-      .filter(value => value.startsWith('text_') || value.startsWith('group_text_'))
-      .map(value => value.replace(/^(text_|group_text_)/, ''));
-    
-    // Handle knowledge changes (user + group)
-    const addedKnowledgeIds = newKnowledgeIds.filter(id => !selectedKnowledgeIds.includes(id));
-    const removedKnowledgeIds = selectedKnowledgeIds.filter(id => !newKnowledgeIds.includes(id));
-    
-    addedKnowledgeIds.forEach(knowledgeId => {
-      // Try to find in user knowledge first, then in group content
-      let selectedItem = availableKnowledge.find(item => item.id === knowledgeId);
-      if (!selectedItem) {
-        selectedItem = allGroupContent.find(item => item.id === knowledgeId);
-      }
-      
-      if (selectedItem) {
-        toggleSelection(knowledgeId);
-        if (onKnowledgeSelection) {
-          onKnowledgeSelection(selectedItem);
-        }
-      }
-    });
-    
-    removedKnowledgeIds.forEach(knowledgeId => {
-      toggleSelection(knowledgeId);
-    });
-    
-    // Handle document changes (user + group)
-    const addedDocumentIds = newDocumentIds.filter(id => !selectedDocumentIds.includes(id));
-    const removedDocumentIds = selectedDocumentIds.filter(id => !newDocumentIds.includes(id));
-    
-    [...addedDocumentIds, ...removedDocumentIds].forEach(documentId => {
-      toggleDocumentSelection(documentId);
-    });
-    
-    // Handle text changes (user + group)
-    const addedTextIds = newTextIds.filter(id => !selectedTextIds.includes(id));
-    const removedTextIds = selectedTextIds.filter(id => !newTextIds.includes(id));
-    
-    [...addedTextIds, ...removedTextIds].forEach(textId => {
-      toggleTextSelection(textId);
-    });
-  }, [selectedKnowledgeIds, selectedDocumentIds, selectedTextIds, availableKnowledge, allGroupContent, toggleSelection, toggleDocumentSelection, toggleTextSelection, onKnowledgeSelection]);
-
-
-  // Note: Document preloading is now handled by useKnowledge hook in parent components
-  // This eliminates the flash of "keine dokumente vorhanden" message
-  // Note: Document selection clearing is handled by store reset in useKnowledge hook
-
-
-  // Note: Selection state is now managed by the store, no local state reset needed
-
-  // Hide component if user is not authenticated
-  if (!user) {
-    return null;
-  }
-  
-  // Only hide component if no functionality is enabled 
-  const hasAnyFeatureEnabled = enableKnowledge || enableDocuments || enableTexts;
-  
-  if (!hasAnyFeatureEnabled) {
-    return null;
-  }
-
-  return (
-    <div className="enhanced-knowledge-selector">
-      <EnhancedSelect
-        label="Hochgeladene Inhalte"
-        inputId="enhanced-knowledge-select"
-        classNamePrefix="enhanced-knowledge-select"
-        className="enhanced-knowledge-select"
-        enableTags={true}
-        enableIcons={true}
-        enableSubtitles={false}
-        isMulti
-        options={knowledgeOptions}
-        // Allow placeholder to hyphenate and wrap: Aus­wählen (soft hyphen)
-        placeholder={"Aus­wählen"}
-        isDisabled={disabled}
-        filterOption={() => true} // Disable default filtering since we handle it
-        onInputChange={(inputValue) => {
-          setCurrentSearchTerm(inputValue);
-        }}
-        value={[
-          // Find selected options from all sources
-          ...selectedKnowledgeIds.map(id =>
-            allKnowledgeOptions.find(option =>
-              option.value === `knowledge_${id}` || option.value === `group_knowledge_${id}`
-            )
-          ).filter(Boolean),
-          ...selectedDocumentIds.map(id =>
-            allKnowledgeOptions.find(option =>
-              option.value === `document_${id}` || option.value === `group_document_${id}`
-            )
-          ).filter(Boolean),
-          ...selectedTextIds.map(id =>
-            allKnowledgeOptions.find(option =>
-              option.value === `text_${id}` || option.value === `group_text_${id}`
-            )
-          ).filter(Boolean)
-        ]}
-        onChange={handleKnowledgeChange}
-        closeMenuOnSelect={false}
-        hideSelectedOptions={true}
-        isClearable={false}
-        isSearchable={true}
-        blurInputOnSelect={true}
-        openMenuOnFocus={false}
-        tabSelectsValue={true}
-        captureMenuScroll={false}
-        menuShouldBlockScroll={false}
-        menuShouldScrollIntoView={false}
-        // Conditional portal: disable when inside dropdown to prevent overlap
-        menuPortalTarget={disableMenuPortal ? null : document.body}
-        menuPosition={disableMenuPortal ? "absolute" : "fixed"}
-        tabIndex={tabIndex}
-        noOptionsMessage={() => {
-          if (currentSearchTerm && currentSearchTerm.trim()) {
-            return `Keine Ergebnisse für "${currentSearchTerm}"`;
-          }
-          if (isLoadingAnyContent) {
-            return 'Lade verfügbare Inhalte...';
-          }
-          if (hasGroupErrors) {
-            return 'Fehler beim Laden einiger Gruppeninhalte';
-          }
-          return 'Keine Inhalte verfügbar. Erstelle Wissen, lade Dokumente hoch oder teile Inhalte mit Gruppen.';
-        }}
-      />
-
-
-      {/* Error status for groups */}
-      {groupContentErrors.length > 0 && (
-        <div className="enhanced-knowledge-selector__errors">
-          Einige Gruppeninhalte konnten nicht geladen werden: {groupContentErrors.map(e => e.groupName).join(', ')}
-        </div>
-      )}
-
-      {/* Show message when no content is available */}
-      {knowledgeOptions.length === 0 && !disabled && !isLoadingAnyContent && (
-        <p className="enhanced-knowledge-selector__no-options">
-          Keine Inhalte verfügbar.<br />
-          Erstelle Wissen in deinem Profil, lade Dokumente hoch, generiere Texte oder teile Inhalte mit Gruppen.
-        </p>
-      )}
-
-      {/* Document Content Extraction Status */}
-      {enableDocuments && selectedDocumentIds.length > 0 && isExtractingDocumentContent && documentExtractionInfo && (
-        <div className={`enhanced-knowledge-selector__extraction-status extraction-status--${documentExtractionInfo.type}`}>
-          <div className="extraction-status__icon">
-            {documentExtractionInfo.type === 'vector_search' && (
-              <div className="extraction-status__spinner"></div>
-            )}
-            {documentExtractionInfo.type === 'success' && '✅'}
-            {documentExtractionInfo.type === 'fallback' && '⚠️'}
-            {documentExtractionInfo.type === 'error' && '❌'}
-          </div>
-          <div className="extraction-status__message">
-            {documentExtractionInfo.message}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-};
-
-EnhancedKnowledgeSelector.propTypes = {
-  onKnowledgeSelection: PropTypes.func,
-  disabled: PropTypes.bool,
-  tabIndex: PropTypes.number,
-  disableMenuPortal: PropTypes.bool
-};
-
-EnhancedKnowledgeSelector.displayName = 'EnhancedKnowledgeSelector';
-
-// Export both components for use in FeatureIcons
-export { ProfileSelector, EnhancedKnowledgeSelector };
